+++ resources/views/topics/topic_detail.blade.php
